+++ src/a-tab-panel.js
@@ -1,3 +1,5 @@
+// @ts-check
+
 const A_TAB_PANEL = 'a-tab-panel';
 const template = document.createElement('template');
 let panelCounter = 0;
@@ -16,18 +18,28 @@ template.innerHTML = /* html */`
 `;
 
 /**
- * `TabPanel` is a panel for a `<a-tab-group>` tab panel.
+ * @summary This is a panel for a `<a-tab-group>` tab panel.
+ * @extends HTMLElement
+ *
+ * @tagname a-tab-panel
+ *
+ * @csspart base - The component's base wrapper.
+ *
+ * @slot - The content of the tab panel.
  */
 class TabPanel extends HTMLElement {
   constructor() {
     super();
 
     if (!this.shadowRoot) {
-      this.attachShadow({ mode: 'open' });
-      this.shadowRoot.appendChild(template.content.cloneNode(true));
+      const shadowRoot = this.attachShadow({ mode: 'open' });
+      shadowRoot.appendChild(template.content.cloneNode(true));
     }
   }
 
+  /**
+   * Lifecycle method that is called when the element is first connected to the DOM.
+   */
   connectedCallback() {
     this.setAttribute('slot', 'panel');
     this.setAttribute('role', 'tabpanel');