+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add default display values to host elements (`a-tab`, `a-tab-panel`).
 - Add `contain: content` to host elements (`a-tab-group`, `a-tab`, `a-tab-panel`) to prevent layout thrashing.
 - Provide better ergonomics for `a-tab` and `a-tab-panel` by automatically setting the `slot` attribute to `tab` and `panel` respectively, if not provided.
+- Generate types declarations for the components.
 
 ## v1.2.0 (2023-10-13)